Fix ProfileEditor "Cancel" button
This patch fixes the strange behaviour of "Cancel" button in ProfileEditor: the
button works only if it is focused when clicking it. This behaviour appeared
after change 62bc0befcf3a1d136428ff42390d7b85f9644038.

Investigation shown that the problem does not appear when textarea onChange
handler is static (useCallback() is used to achieve this) and uses
setFieldValue() instead of field's setValue(). Some internal Formik magic works
here. Don't try to understand, just leave it as it is.