Conflict between savingDraft and payloadToKey safeguards
If COMPOSE_DRAFT_SAVE is sent, but blocked by payloadToKey condition,
state.compose.savingDraft is not set back to false. Because of
that, ComposeDraftSaver will not send further COMPOSE_DRAFT_SAVE
actions.

Actually, payloadToKey safeguard is not needed here, because
COMPOSE_DRAFT_SAVE will not be sent too often.

The same change is made for the comment draft saver.