Prevent adding invalid URLs to the history stack
When a user opens a post on a foreign node and then moves to one of
the home pages (News, Explore, Instants, etc.), the location
'/post/null' is added to the history stack.

It happens because INIT_FROM_LOCATION clears large parts of the state,
including state.detailedposting. Usually GO_TO_PAGE sets a new valid
location immediately, but GO_HOME_PAGE postpones this till finishing
its saga and the current invalid location is added to the history
stack.

This patch is a dirty hack just to prevent the error. But the current
navigation code is overly complex and fragile and should be rewritten.